requirements.txt
pydantic
llama-cpp-python